+++ packages/admin-rte/src/core/Rte.tsx
@@ -12,12 +12,13 @@ import {
 import * as React from "react";
 
 import Controls from "./Controls";
+import defaultBlocktypeMap, { mergeBlocktypeMaps } from "./defaultBlocktypeMap";
 import composeFilterEditorFns from "./filterEditor/composeFilterEditorFns";
 import defaultFilterEditorStateBeforeUpdate from "./filterEditor/default";
 import manageDefaultBlockType from "./filterEditor/manageStandardBlockType";
 import removeBlocksExceedingBlockLimit from "./filterEditor/removeBlocksExceedingBlockLimit";
 import * as sc from "./Rte.sc";
-import { ICustomBlockTypeMap, ToolbarButtonComponent } from "./types";
+import { IBlocktypeMap, ICustomBlockTypeMap_Deprecated, ToolbarButtonComponent } from "./types";
 import createBlockRenderMap from "./utils/createBlockRenderMap";
 
 const mandatoryFilterEditorStateFn = composeFilterEditorFns([removeBlocksExceedingBlockLimit, manageDefaultBlockType]);
@@ -45,7 +46,7 @@ export type SupportedThings =
 export interface IRteOptions {
     supports: SupportedThings[];
     listLevelMax: number;
-    customBlockMap?: ICustomBlockTypeMap;
+    blocktypeMap: IBlocktypeMap;
     overwriteLinkButton?: ToolbarButtonComponent;
     overwriteLinksRemoveButton?: ToolbarButtonComponent;
     customToolbarButtons?: ToolbarButtonComponent[];
@@ -58,6 +59,9 @@ export interface IRteOptions {
     filterEditorStateBeforeUpdate?: FilterEditorStateBeforeUpdateFn;
     maxBlocks?: number;
     standardBlockType: DraftBlockType;
+
+    // @deprecated
+    customBlockMap?: ICustomBlockTypeMap_Deprecated;
 }
 
 export type IOptions = Partial<IRteOptions>;
@@ -96,6 +100,7 @@ const defaultOptions: IRteOptions = {
     // standardBlockType can be set to any supported block-type,
     // when set to something other than "unstyled" the unstyled-blockType is disabled (does not show up in the Dropdown)
     standardBlockType: "unstyled",
+    blocktypeMap: defaultBlocktypeMap,
 };
 
 export interface IRteRef {
@@ -117,7 +122,22 @@ const Rte: React.RefForwardingComponent<any, IProps> = (props, ref) => {
     const { value: editorState, onChange, options: passedOptions } = props;
     const editorRef = React.useRef<DraftJsEditor>(null);
     const editorWrapperRef = React.useRef<HTMLDivElement>(null);
-    const options = passedOptions ? { ...defaultOptions, ...passedOptions } : defaultOptions; // merge default options with passed options
+
+    // merge default options with passed options
+    let options = passedOptions ? { ...defaultOptions, ...passedOptions } : defaultOptions;
+
+    // extract deprecated options and handle them specially
+    let deprecatedCustomBlockMap: ICustomBlockTypeMap_Deprecated = {};
+    if (options.customBlockMap) {
+        deprecatedCustomBlockMap = options.customBlockMap;
+        delete options.customBlockMap;
+    }
+
+    // blocktypes need an extra merge as they have their own merge strategy
+    options = {
+        ...options,
+        blocktypeMap: mergeBlocktypeMaps(defaultBlocktypeMap, deprecatedCustomBlockMap, options.blocktypeMap),
+    };
 
     /**
      * Expose methods
@@ -154,7 +174,7 @@ const Rte: React.RefForwardingComponent<any, IProps> = (props, ref) => {
         [filterEditorStateBeforeUpdate, supports, listLevelMax, maxBlocks, standardBlockType, onChange],
     );
 
-    const blockRenderMap = createBlockRenderMap({ customBlockTypeMap: options.customBlockMap });
+    const blockRenderMap = createBlockRenderMap({ blocktypeMap: options.blocktypeMap });
 
     function handleKeyCommand(command: DraftEditorCommand) {
         const commandToSupportsMap: Partial<Record<DraftEditorCommand, SupportedThings>> = {