+++ packages/admin-rte/src/field/createFinalFormRte.tsx
@@ -57,7 +57,8 @@ function createFinalFormRte<T = any>(config: IConfig<T> = defaultConfig) {
                 value={createStateFromRawContent(content)}
                 plainTextOnly={plainTextOnly}
                 options={{
-                    customBlockMap: rteOptions ? rteOptions.customBlockMap : undefined,
+                    blocktypeMap: rteOptions ? rteOptions.blocktypeMap : undefined,
+                    customBlockMap: rteOptions ? rteOptions.customBlockMap : undefined, // deprecated
                 }}
             />
         );