+++ packages/admin-stories/src/admin-rte/RteAllOptions.tsx
@@ -1,4 +1,5 @@
 import { IMakeRteApiProps, IRteApiProps, IRteOptions, IRteRef, LinkDecorator, makeRteApi, Rte } from "@comet/admin-rte";
+import { Typography } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
 import { convertFromRaw, convertToRaw } from "draft-js";
 import * as React from "react";
@@ -25,7 +26,13 @@ export const apiOptions: IRteApiProps<ContentFormat> = {
     debounceDelay: 400, // delay when onDebouncedContentChange after editor content changed
 };
 
-const GreenCustomHeader: React.FC = ({ children }) => <span style={{ color: "green" }}>{children}</span>;
+const RedCustomHeader: React.FC = ({ children }) => (
+    <Typography variant="h1" style={{ color: "red" }}>
+        {children}
+    </Typography>
+);
+const GreenCustomHeader: React.FC = ({ children }) => <h2 style={{ color: "green" }}>{children}</h2>;
+
 export const rteOptions: IRteOptions = {
     supports: [
         "bold",
@@ -48,13 +55,29 @@ export const rteOptions: IRteOptions = {
         "links-remove",
     ],
     listLevelMax: 2,
-    customBlockMap: {
+    blocktypeMap: {
+        // overwrite built-in blocktypes
+        "header-one": {
+            label: "HEADING 1",
+            // more on this setting here:
+            // https://draftjs.org/docs/advanced-topics-custom-block-render-map/#configuring-block-render-map
+            renderConfig: {
+                element: (p) => <RedCustomHeader {...p} />,
+                aliasedElements: ["h1"],
+            },
+        },
+        // define new blocktypes
         "header-custom-green": {
             label: "Custom Green Header",
             renderConfig: {
-                element: "h1",
-                wrapper: <GreenCustomHeader />,
+                element: (p) => <GreenCustomHeader {...p} />,
+                aliasedElements: ["h2"],
             },
+        },
+
+        "ordered-list-item": {
+            // this is funky: moves the ordered-list-item to the dropdown (not recommended to change the default behaviour)
+            group: "dropdown",
         },
     },
     draftJsProps: {