+++ packages/admin-rte/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [next]
+
+### Highlights
+
+-   Add default styles (MUI) to built-in blocktypes
+-   Make built-in blocktypes styleable
+
+### Internal Changes
+
+-   Rename prop-name "customBlockMap" to "blocktypeMap", deprecate prop-name "customBlockMap"
 ## [1.1.1] - 12. Jan 2021 - re-release under new name
 
 This package has been renamed to @comet/admin-rte